requirements.txt
flask
psycopg2-binary
pyTelegramBotAPI
requests
gunicorn